+++ backend/src/main/java/pl/freeedu/backend/lesson/repository/GroupHasLessonRepository.java
@@ -8,16 +8,16 @@ import pl.freeedu.backend.lesson.model.GroupHasLesson;
 import java.util.List;
 
 @Repository
-public interface GroupHasLessonRepository extends JpaRepository<GroupHasLesson, Long> {
+public interface GroupHasLessonRepository extends JpaRepository<GroupHasLesson, Integer> {
 
-    List<GroupHasLesson> findByLessonId(Long lessonId);
+    List<GroupHasLesson> findByLessonId(Integer lessonId);
 
-    void deleteByLessonId(Long lessonId);
+    void deleteByLessonId(Integer lessonId);
 
     @Query("SELECT new pl.freeedu.backend.lesson.dto.GroupDto(g.id, g.name) " +
            "FROM pl.freeedu.backend.usergroup.model.UserGroup g JOIN pl.freeedu.backend.lesson.model.GroupHasLesson ghl ON g.id = ghl.groupId " +
            "WHERE ghl.lessonId = :lessonId")
-    List<pl.freeedu.backend.lesson.dto.GroupDto> findGroupsForLesson(Long lessonId);
+    List<pl.freeedu.backend.lesson.dto.GroupDto> findGroupsForLesson(Integer lessonId);
 
     @Query("SELECT ghl.lessonId FROM pl.freeedu.backend.lesson.model.GroupHasLesson ghl WHERE ghl.groupId = :groupId")
     List<Long> findLessonIdsByGroupId(Integer groupId);
